# src/main/resources/api-specification.yaml
openapi: 3.0.3
info:
  title: Sistema de Conversão de Moedas do Reino SRM
  description: |
    API para conversão de moedas entre Ouro Real (OR) e Tibar (TB) no Mercado de Pulgas dos Mil Saberes.
    Funcionalidades:
      - Conversão entre Ouro Real (OR) e Tibar (TB)
      - Multiplicadores por categoria de produto
      - Histórico completo de transações
      - Taxas de câmbio atualizáveis
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor de Desenvolvimento
  - url: https://api.srm.kingdom
    description: Servidor de Produção
...
paths:
...
  /products:
    get:
      tags:
        - Produtos
      summary: Listar todos os produtos
      operationId: getAllProducts
      responses:
        '200':
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductSimple'
    post:
      tags:
        - Produtos
      summary: Criar um novo produto
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Produto criado
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/{id}:
    get:
      tags:
        - Produtos
      summary: Obter produto por id
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSimple'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /transactions:
    get:
      tags:
        - Transações
      summary: Listar transações (filtros simples)
      operationId: listTransactions
      responses:
        '200':
          description: Página de transações
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
    post:
      tags:
        - Transações
      summary: Criar uma transação
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreate'
      responses:
        '201':
          description: Transação criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ProductSimple:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Pele de Urso das Montanhas"
        description:
          type: string
          example: "Pele resistente de urso das montanhas geladas"
        categoryName:
          type: string
          enum: [Peles, Madeira, Hidromel]
          example: "Peles"
        categoryMultiplier:
          type: number
          format: decimal
          example: 1.0000
        basePrice:
          type: number
          format: decimal
          example: 50.0000
        baseCurrencyCode:
          type: string
          example: "OR"
        baseCurrencySymbol:
          type: string
          example: "⚜"
        kingdomName:
          type: string
          example: "Montanhas dos Anões"
    ProductCreate:
      type: object
      required:
        - name
        - categoryName
        - basePrice
        - baseCurrencyCode
      properties:
        name:
          type: string
        description:
          type: string
        categoryName:
          type: string
        basePrice:
          type: number
          format: decimal
        baseCurrencyCode:
          type: string
        kingdomName:
          type: string
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
          format: int64
        originalAmount:
          type: number
          format: decimal
        convertedAmount:
          type: number
          format: decimal
        exchangeRate:
          type: number
          format: decimal
        transactionDate:
          type: string
          format: date-time
    TransactionCreate:
      type: object
      required:
        - transactionType
        - originalAmount
        - convertedAmount
        - exchangeRate
      properties:
        transactionType:
          type: string
          enum: [EXCHANGE, PURCHASE, SALE]
        originalAmount:
          type: number
          format: decimal
        convertedAmount:
          type: number
          format: decimal
        exchangeRate:
          type: number
          format: decimal
        productId:
          type: integer
          format: int64
          nullable: true
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2025-01-30T01:32:15.123456"
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Resource Not Found"
        message:
          type: string
          example: "Currency not found: INVALID"
        validationErrors:
          type: object
          additionalProperties:
            type: string
          nullable: true